Fix horizontal speed
